refactor: unified NexusNode tree, Annotator, FileValidator, NomadParser
Phase 1-3 of the Unified NeXus File Handler refactor:

- Move NexusNode tree to nexus/nexus_tree.py; dataconverter/nexus_tree.py
  becomes a re-export shim for backward compatibility
- Add nexus/annotation.py: Annotator class with NexusTree-based schema
  resolution replacing get_inherited_hdf_nodes / get_nxdl_doc chains;
  structured inheritance display with per-element concept paths
  (NXinstrument::/energy, NXarpes::/NXentry/NXinstrument/analyser/lens_mode)
- Add nexus/handler.py: NexusFileHandler + NexusVisitor protocol stubs
  (foundation for Phase 4 full handler)
- Refactor dataconverter/validation.py: FileValidator class with instance-level
  node cache replaces visititems closures; validate_hdf_group_against is a
  1-line wrapper
- Add NexusNode helpers: concept_path, get_inheritance_docs,
  get_inheritance_enums, get_inheritance_concept_paths,
  has_nxcollection_parent, best_child_for, find_node_at_path
- Add _xml_path_in_nxdl module helper for precise within-NXDL element paths
- nomad/parser.py: NomadParser thin wrapper with TODO skeleton for
  future NomadVisitor migration
- 329 tests passing
