+++ .cspell/custom-dictionary.txt
@@ -234,6 +234,8 @@ showlegend
 signedinteger
 skiprows
 straße
+subelems
+sublists
 submoduled
 superproject
 threed
